Add get_etl_deps DAG
I suspect ETL DAGs are failing because there's an import
timeout that results from the processing time it takes to
query the DB and re-parse the dependencies and their operators.
I'll try creating a once-a-day static csv for the ETL DAGs to read
instead of live querying, and if that doesn't work I'll try
storing it in the Airflow metadata DB because that shouldn't
affect the connection pooling with the prod and QA DBs.