Restructure querying of ETL dependencies
I suspected that when the scheduler filled up the dagbag that the computation time
and resources required to connect to the database, query the dependencies, and then
process them was causing a timeout and as a result Celery tasks were failing. I was right.
Now the master ETL DAGs wait for an external DAG to query the dependencies and save them
down as a csv that can be accessed by any of the workers and scheduler, as long as they're
not on the same node as the location of the volume whose driver is vieux/sshfs, which uses
ssh to allow all the other nodes to see the same folder. This prevents the need to save down
a copy of the dependencies on each node, and the effort required to update it.

Additionally, I renamed the DAGs and their .py files to have imperative descriptions,
in the same way that Python functions should be named, in order to avoid ambiguity and
to be more precise.