Restructure get_etl_deps for efficiency
Created a view in the DB in order not to have to perform a complex
SQL query every time the scheduler heartbeats. Also this way the SQL
is stored in the DB instead of the DAG file.

Created several .json files to read in instead of performing the computation
for every heartbeat.

Added functionality for updating a view reference in DB as well.

Closes #16